load_pdfs_and_collect_metadata/requirements.txt
#pip install 
pdfplumber
mysql-connector-python